requirements.txt
Flask
Flask-RESTful
Flask-Bootstrap
google-api-python-client
google-auth
google-auth-oauthlib
gspread
gunicorn
itsdangerous
Jinja2
oauthlib
peewee
slackclient
Werkzeug
WTForms
